+++ cli/impl/cliRecordsDelete.go
@@ -1,6 +1,7 @@
 package impl
 
 import (
+	"encoding/json"
 	"ffsyncclient/cli"
 	"ffsyncclient/consts"
 	"ffsyncclient/fferr"
@@ -11,10 +12,13 @@ import (
 type CLIArgumentsRecordsDelete struct {
 	Collection string
 	RecordID   string
+	SoftDelete bool
 }
 
 func NewCLIArgumentsRecordsDelete() *CLIArgumentsRecordsDelete {
-	return &CLIArgumentsRecordsDelete{}
+	return &CLIArgumentsRecordsDelete{
+		SoftDelete: false,
+	}
 }
 
 func (a *CLIArgumentsRecordsDelete) Mode() cli.Mode {
@@ -27,7 +31,7 @@ func (a *CLIArgumentsRecordsDelete) PositionArgCount() (*int, *int) {
 
 func (a *CLIArgumentsRecordsDelete) ShortHelp() [][]string {
 	return [][]string{
-		{"ffsclient delete <collection> <record-id>", "Delete the specified record"},
+		{"ffsclient delete <collection> <record-id> [--soft]", "Delete the specified record"},
 	}
 }
 
@@ -36,6 +40,7 @@ func (a *CLIArgumentsRecordsDelete) FullHelp() []string {
 		"$> ffsclient delete <collection> <record-id>",
 		"",
 		"Delete the specific record from the server",
+		"If --soft is specified we do not really delete the record, but only add {deleted:true} to its payload",
 	}
 }
 
@@ -44,6 +49,10 @@ func (a *CLIArgumentsRecordsDelete) Init(positionalArgs []string, optionArgs []c
 	a.RecordID = positionalArgs[1]
 
 	for _, arg := range optionArgs {
+		if arg.Key == "soft" && arg.Value == nil {
+			a.SoftDelete = true
+			continue
+		}
 		return fferr.DirectOutput.New("Unknown argument: " + arg.Key)
 	}
 
@@ -55,6 +64,7 @@ func (a *CLIArgumentsRecordsDelete) Execute(ctx *cli.FFSContext) int {
 	ctx.PrintVerbose("")
 	ctx.PrintVerboseKV("Collection", a.Collection)
 	ctx.PrintVerboseKV("RecordID", a.RecordID)
+	ctx.PrintVerboseKV("Soft", a.SoftDelete)
 
 	// ========================================================================
 
@@ -89,10 +99,48 @@ func (a *CLIArgumentsRecordsDelete) Execute(ctx *cli.FFSContext) int {
 
 	// ========================================================================
 
-	err = client.DeleteRecord(ctx, session, a.Collection, a.RecordID)
-	if err != nil {
-		ctx.PrintFatalError(err)
-		return consts.ExitcodeError
+	if a.SoftDelete {
+
+		record, err := client.GetRecord(ctx, session, a.Collection, a.RecordID, true)
+		if err != nil {
+			ctx.PrintFatalError(err)
+			return consts.ExitcodeError
+		}
+
+		var jsonpayload langext.H
+		err = json.Unmarshal(record.DecodedData, &jsonpayload)
+		if err != nil {
+			ctx.PrintFatalError(fferr.DirectOutput.Wrap(err, "failed to unmarshal"))
+			return consts.ExitcodeError
+		}
+		jsonpayload["deleted"] = true
+
+		plainpayload, err := json.Marshal(jsonpayload)
+		if err != nil {
+			ctx.PrintFatalError(fferr.DirectOutput.Wrap(err, "failed to re-marshal payload"))
+			return consts.ExitcodeError
+		}
+
+		payload, err := client.EncryptPayload(ctx, session, a.Collection, string(plainpayload))
+		if err != nil {
+			ctx.PrintFatalError(err)
+			return consts.ExitcodeError
+		}
+
+		err = client.PutRecord(ctx, session, a.Collection, a.RecordID, payload, false, false)
+		if err != nil {
+			ctx.PrintFatalError(err)
+			return consts.ExitcodeError
+		}
+
+	} else {
+
+		err = client.DeleteRecord(ctx, session, a.Collection, a.RecordID)
+		if err != nil {
+			ctx.PrintFatalError(err)
+			return consts.ExitcodeError
+		}
+
 	}
 
 	// ========================================================================
